chore(auth): added loader module
